==> notebooks/0.0-first-experiment.ipynb <==
!pip install transformers

# Load model directly
from transformers import AutoTokenizer, AutoModelForSequenceClassification,RobertaModel, RobertaForSequenceClassification

tokenizer = AutoTokenizer.from_pretrained("SamLowe/roberta-base-go_emotions")
model = RobertaModel.from_pretrained("SamLowe/roberta-base-go_emotions", add_pooling_layer = False)

inp = tokenizer("I have bad day", return_tensors="pt")
out1 = model(**inp)[0][:, 0, :]

inp = tokenizer("I have a great day", return_tensors="pt")
out2 = model(**inp)[0][:, 0, :]

from torch.nn import functional as F

F.cosine_similarity(out1, out2)

from transformers import AutoTokenizer, RobertaModel


tokenizer = AutoTokenizer.from_pretrained("cardiffnlp/twitter-roberta-base-emotion-multilabel-latest")
model = RobertaModel.from_pretrained("cardiffnlp/twitter-roberta-base-emotion-multilabel-latest", add_pooling_layer = False)

inp = tokenizer("I have bad day", return_tensors="pt")
out1 = model(**inp)[0][:, 0, :]

inp = tokenizer("I have a great day", return_tensors="pt")
out2 = model(**inp)[0][:, 0, :]

from torch.nn import functional as F


F.cosine_similarity(out1, out2)

F.pdist()

(1 -0.14) / 2

tokenizer = AutoTokenizer.from_pretrained("roberta-base")
model = RobertaModel.from_pretrained("roberta-base", add_pooling_layer = False)

inp = tokenizer("I have bad day", return_tensors="pt")
out1 = model(**inp)[0][:, 0, :]

inp = tokenizer("I have a great day", return_tensors="pt")
out2 = model(**inp)[0][:, 0, :]

from torch.nn import functional as F

F.cosine_similarity(out1, out2)
