==> notebooks/1.0-text-data-filtering.ipynb <==
# Data Filtering of Emotions Datasets

Datasets:

1. [go_emotions](https://huggingface.co/datasets/go_emotions) — 58k carefully curated Reddit comments labeled for 27 emotion categories or Neutral.
2. [sem_eval_2018_task_1](https://huggingface.co/datasets/sem_eval_2018_task_1) — a dataset of Twitter tweets in Arabic, English (11k entries), Spanish to automatically determine the intensity of emotions (E) and intensity of sentiment (aka valence V) of the tweeters from their tweets.

Both datasets are designed to solve multi-label emotion classification task.

## Filtering Idea

Idea is to choose out of these datasets those texts that are somehow related to the possible answer on such questions like:
- How was your day?
- How do you feel today?
- How its going?
- What's up?
- How do you do?
- What can you say about your day?
- Describe your day?

In other words, choose only diary-style texts.

It will help to filer out completely unrelated and trash texts that will not be suitable for out main task goal — recommend a quote for a user that responeded with a description of their day (like a post from dairy).

from datasets import load_dataset


go_emotions = load_dataset('go_emotions', 'raw')
sem_eval = load_dataset('sem_eval_2018_task_1', 'subtask5.english')

go_emotions

Remove unnecessary columns

go = go_emotions['train']

def remove_columns_go(ds):
    for col in ['id', 'author', 'subreddit', 'link_id', 'parent_id', 'created_utc', 'rater_id']:
        ds.pop(col)
    return ds

go = go.map(remove_columns_go)

go

sum(go['neutral']) / go.num_rows

26% of text are completely neutral, so almost useless for us.

sem_eval

### Idea — create a list of keywords to search in texts using ChatGPT

We have asked ChatGPT the following:

```
Help me to create a key words list that can be related to texts that are written to answer the following questions:
- How was your day?
- How do you feel today?
- How its going?
- What's up?
- How do you do?
- What can you say about your day?
- Describe your day?

I need the keywords that describe a possible text that answers the questions I provided.
Take a deep breath and answer in the format of python list: ['word1', 'word2'].
```

Below we have a list that model wrote for us.

keywords = set(['good', 'great', 'fine', 'wonderful', 'fantastic', 'awesome', 'terrible', 'bad', 'okay', 'alright', 'not', 'much', 'not much', 'so-so', 'just', 'hanging', 'in', 'there', 'not', 'too', 'shabby', 'feeling', 'positive', 'negative', 'today', 'now', 'not', 'much', 'not', 'a', 'lot', 'just', 'working', 'busy', 'challenging', 'eventful', 'nothing', 'special', 'ordinary', 'usual', 'remarkable', 'pleasant', 'difficult', 'stressful', 'exciting', 'calm', 'quiet', 'productive', 'relaxing', 'unproductive', 'tired', 'energetic', 'happy', 'sad', 'content', 'stressed', 'anxious', 'bored', 'frustrated', 'confused', 'positive', 'negative', 'neutral', 'normal', 'extraordinary', 'routine', 'typical', 'unusual', 'nothing', 'much', 'out', 'the', 'ordinary', 'just', 'the', 'same', 'as', 'usual', 'not', 'a', 'lot', 'to', 'report', 'same', 'old', 'same', 'old', 'nothing', 'special', 'went', 'well', 'went', 'bad', 'went', 'fine', 'went', 'poorly', 'went', 'great', 'went', 'horrible', 'went', 'so-so', 'went', 'as', 'expected', 'went', 'unexpectedly', 'not', 'much', 'to', 'tell', 'interesting', 'exciting', 'ordinary', 'mundane', 'text', 'write', 'response', 'reply', 'communicate', 'communicative', 'express', 'yourself', 'convey', 'your', 'thoughts', 'elaborate', 'tell', 'me', 'about', 'it', 'inform', 'me', 'informative', 'insightful', 'detail', 'thorough', 'brief', 'concise', 'informal', 'formal', 'vivid', 'colorful', 'dull', 'monotonous', 'typical', 'atypical', 'daily', 'usual', 'extraordinary', 'usual', 'routine', 'daily', 'routine', 'ordinary', 'extraordinary', 'common', 'uncommon', 'daily', 'activities', 'daily', 'events', 'daily', 'occurrences', 'daily', 'experiences'])

len(keywords)

Let's remove stopwords from this list

from nltk.corpus import stopwords

kw = set([k for k in keywords if k not in stopwords.words('english')])

len(kw)

Now, trying to filter this out

filtered_samples = []

for text in go['text'][:1000]:
    for k in kw:
        if k in text.lower():
            filtered_samples.append(text)
            break

filtered_samples, len(filtered_samples)

filtered_samples = []

for text in sem_eval['train']['Tweet'][:1000]:
    for k in kw:
        if k in text.lower():
            filtered_samples.append(text)
            break

filtered_samples, len(filtered_samples)

Even though, using this idea we filter around 70% out of 1000 samples from both dataset, the filtered ones are still not quite relevant and they do not look as a texts that answer defined questions. So we need something different. 

### Idea — ask LLMs to say whether the text are close to diary-style or not

from langchain.llms import Ollama


# Make sure you do these steps before running it
# 1. https://ollama.ai/download
# 2. ollama serve
# 3. ollama pull mistral
llm = Ollama(model="mistral")

zipped = zip(
    ['My day was great! I have done all the things I planned to do. Also, I met my brother and it was pleasure to see him after 5 years he moved to USA.',
     'Thank you for reporting the issue! We wil think how to approach this problem.',
     'You are not right here! Go away from this chat.'],
     ['+', '-', '-'])
examples = []
for ref, trn in zipped:
    examples.append({
        'text': ref,
        'answer': trn
    })
examples

from langchain import PromptTemplate, FewShotPromptTemplate, LLMChain


example_template = """
Text: "{text}"
Is above text written in a diary-style? Answer: {answer}
"""

example_prompt = PromptTemplate(
   input_variables=["text", "answer"],
   template=example_template
)

prefix = """
Your task is to decide is the provided text written in a diary-style? Answer ONLY with "+" or "-" as an output!
DO NOT output such phrases as "I understand your task", "I am an AI language model" or something similar.
It COULD NOT be no response, ONLY "+" or "-". DO NOT ASK for additional context information or few more examples!\n
"""
suffix = """
Text: "{text}"
Is above text written in a diary-style? Answer ONLY with "+" or "-": 
"""

few_shot_prompt_template = FewShotPromptTemplate(
   examples=examples,
   example_prompt=example_prompt,
   prefix=prefix,
   suffix=suffix,
   input_variables=["text"],
   example_separator="\n\n"
)

fs_llm_chain = LLMChain(
   prompt=few_shot_prompt_template,
   llm=llm
)

for text in go['text'][:10]:
    print(f'TEXT: {text}')
    llm_output = fs_llm_chain.run(few_shot_prompt_template.format(text=text))
    print(f'PRED: {llm_output}')
    print('------------------------')

This approach requires the postprocessing of the output. We can't predict all LLM's outputs, so there is always a risk to get bad answer. The LLMs also tend to hallucinations and require more time for better engineered prompt.

# Conclusion

Let's try to find more datasets, maybe not specifically with emotions but with dairy-style texts. Another idea is to keep these emotions datasets since the data domain is close and we can still fine-tune our future model.
