==> notebooks/1.1-text-data-preprocessing.ipynb <==
# Data Preprocessing of Text Datasets

## 1. Diary-style Text Datasets 

Datasets:

1. [journal-entries-with-labelled-emotions](https://www.kaggle.com/datasets/madhavmalhotra/journal-entries-with-labelled-emotions) — 1500 journal entries labelled with 18 emotions from people reflecting on their day.
2. [Diary-Entry-To-Rap](https://huggingface.co/datasets/chaudha7/Diary-Entry-To-Rap) — 175 diary entries with age and gender writer's info plus transformed diary into rap text with certain mood and style.

From both datasets only texts will be taken.

### Read only diary-style text entries

import pandas as pd


journals_ds_dir = '../data/journals.csv'

journals_df = pd.read_csv(journals_ds_dir)
journals_df.head(1)

journals = journals_df['Answer'].to_list()
len(journals), journals[:5]

from datasets import load_dataset


diaries_ds = load_dataset('chaudha7/Diary-Entry-To-Rap')
diaries_ds

diaries = diaries_ds['train']['Diary Entry']
len(diaries), diaries[:5]

### Prepare for labeling

diary_texts = journals + diaries
len(diary_texts)

### Labeling

from transformers import pipeline


reddit_classifier = pipeline(task="text-classification", model="SamLowe/roberta-base-go_emotions", top_k=None, padding=True, truncation=True)
twitter_classifier = pipeline(task="text-classification", model="cardiffnlp/twitter-roberta-base-emotion-multilabel-latest", top_k=None, padding=True, truncation=True)

reddit_outputs = reddit_classifier(diary_texts)
reddit_outputs[:5]

import pickle


with open('../data/reddit_outputs.pickle', 'wb') as handle:
    pickle.dump(reddit_outputs, handle)
# with open('../data/reddit_outputs.pickle', 'rb') as handle:
    # reddit_outputs = pickle.load(handle)

twitter_outputs = twitter_classifier(diary_texts)
twitter_outputs[:5]

import pickle


with open('../data/twitter_outputs.pickle', 'wb') as handle:
    pickle.dump(twitter_outputs, handle)
# with open('../data/twitter_outputs.pickle', 'rb') as handle:
    # twitter_outputs = pickle.load(handle)

data = []

for i, text in enumerate(diary_texts):
    labels = reddit_outputs[i]
    data.append({"Text": text, "Labels": labels})

df = pd.DataFrame(data)
df = pd.concat([df.drop(['Labels'], axis=1), df['Labels'].apply(lambda x: pd.Series({d['label']: d['score'] for d in x}))], axis=1)
df.to_csv('../data/diaries_labeled_reddit.csv')
df.head()

df.describe()

data = []

for i, text in enumerate(diary_texts):
    labels = twitter_outputs[i]
    data.append({"Text": text, "Labels": labels})

df = pd.DataFrame(data)
df = pd.concat([df.drop(['Labels'], axis=1), df['Labels'].apply(lambda x: pd.Series({d['label']: d['score'] for d in x}))], axis=1)
df.to_csv('../data/diaries_labeled_twitter.csv')
df.head()

df.describe()

## 2. Reddit and Twitter Posts Datasets

Datasets:

1. [go_emotions](https://huggingface.co/datasets/go_emotions) — 58k carefully curated Reddit comments labeled for 27 emotion categories or Neutral.
2. [sem_eval_2018_task_1](https://huggingface.co/datasets/sem_eval_2018_task_1) — a dataset of Twitter tweets in Arabic, English (11k entries), Spanish to automatically determine the intensity of emotions (E) and intensity of sentiment (aka valence V) of the tweeters from their tweets.

Both datasets are designed to solve multi-label emotion classification task. From both datasets only texts will be taken.

TODO: try
