==> notebooks/1.2-quotes-data-preprocessing.ipynb <==
# Data Preprocessing of Quotes Datasets

## 1. Quotes Dataset From Kaggle

[Quotes Dataset](https://www.kaggle.com/datasets/hiteshsuthar101/quotes-dataset) — 852 quotes entries parsed and labeled from [quotepark.com](quotepark.com) 

import pandas as pd


quotes_ds_dir = '../data/quotes.csv'

df = pd.read_csv(quotes_ds_dir)
df.head(1)

### 0.0 Labeling using classifiers

quotes = df['Quote'].to_list()
quotes[:5]

from transformers import pipeline


reddit_classifier = pipeline(task="text-classification", model="SamLowe/roberta-base-go_emotions", top_k=None, padding=True, truncation=True)
twitter_classifier = pipeline(task="text-classification", model="cardiffnlp/twitter-roberta-base-emotion-multilabel-latest", top_k=None, padding=True, truncation=True)

import pickle

reddit_outputs = reddit_classifier(quotes)
with open('../data/quotes_reddit_outputs.pickle', 'wb') as handle:
    pickle.dump(reddit_outputs, handle)

import pickle

twitter_outputs = twitter_classifier(quotes)
with open('../data/quotes_twitter_outputs.pickle', 'wb') as handle:
    pickle.dump(twitter_outputs, handle)

data = []

for i, text in enumerate(quotes):
    labels = reddit_outputs[i]
    data.append({"Text": text, "Labels": labels})

df = pd.DataFrame(data)
df = pd.concat([df.drop(['Labels'], axis=1), df['Labels'].apply(lambda x: pd.Series({d['label']: d['score'] for d in x}))], axis=1)
df.to_csv('../data/quotes_labeled_reddit.csv')
df.head()

data = []

for i, text in enumerate(quotes):
    labels = twitter_outputs[i]
    data.append({"Text": text, "Labels": labels})

df = pd.DataFrame(data)
df = pd.concat([df.drop(['Labels'], axis=1), df['Labels'].apply(lambda x: pd.Series({d['label']: d['score'] for d in x}))], axis=1)
df.to_csv('../data/quotes_labeled_twitter.csv')
df.head()

### 1.1. Dataset preprocessing
We have to remove the columns that we don't need for our task and split the tags column into a set of tags.

df = df[['Quote', 'Tags']]
df["Tags"] = df["Tags"].apply(lambda x: {e.strip().lower() for e in x.split(",")} if isinstance(x, str) else [])

### 1.2. Add tags from the GoEmotions dataset to the quotes dataset
It is necessary to add tags from the GoEmotions dataset to the quotes dataset, because we will use the model fine-tuned on GoEmotions dataset.

Here we will use the model [SamLowe/roberta-base-go_emotions](https://huggingface.co/SamLowe/roberta-base-go_emotions) model.

#### 1.2.1. Use RoBERTa model to predict tags on quote text
Get the tags from the model (with threshold = 0.2) and filter them using cleaned tag set.

from transformers import pipeline
pipe = pipeline("text-classification", model="SamLowe/roberta-base-go_emotions")

# Filtered tag set for quotes from the go emotions
my_set =  {'admiration',
           'amusement',
           'anger',
           'annoyance',
           'confusion',
           'curiosity',
           'desire',
           'disappointment',
           'disgust',
           'embarrassment',
           'excitement',
           'fear',
           'gratitude',
           'grief',
           'joy',
           'love',
           'nervousness',
           'optimism',
           'pride',
           'realization',
           'relief',
           'remorse',
           'sadness',
           'surprise'}

tags = pipe(list(df["Quote"]), truncation=True, top_k=None)

df["model_tags"] = [{tag["label"] for tag in quote if tag["score"] > 0.2 and tag["label"] in my_set} for quote in tags]

#### 1.2.2. Add tags from the quotes dataset intersection with the GoEmotions dataset

df["ds_tags"] = df["Tags"].apply(lambda x: my_set.intersection(x))

df

#### 1.2.3. Use `fasttext` embeddings to find similar tags

import fasttext
from sklearn.metrics.pairwise import cosine_similarity


ft = fasttext.load_model('cc.en.300.bin')

my_set_embed = [ft.get_word_vector(e) for e in my_set]

from tqdm.notebook import tqdm


all_ds_tags = set(df["Tags"].explode().value_counts().keys())
tag_to_tag = {}

for e in tqdm(all_ds_tags):
    for w, emb in zip(my_set, my_set_embed):
        if cosine_similarity([ft.get_word_vector(e)], [emb])[0][0] > 0.5:
            tag_to_tag[e] = w

Clean pairs of tags by hand.

tag_to_tag

cleaned_tag2tag = {'encouragement': 'gratitude',
 'danger': 'fear',
 'intention': 'desire',
 'regret': 'sadness',
 'disappointment': 'sadness',
 'rage': 'disgust',
 'appetite': 'desire',
 'agony': 'sadness',
 'optimist': 'optimism',
 'kindness': 'gratitude',
 'pleasure': 'joy',
 'pride': 'joy',
 'imagination': 'curiosity',
 'amazement': 'disgust',
 'awe': 'admiration',
 'happiness': 'optimism',
 'acknowledgment': 'gratitude',
 'loneliness': 'sadness',
 'delight': 'joy',
 'admire': 'admiration',
 'sorrow': 'sadness',
 'laughter': 'joy',
 'hatred': 'disgust',
 'want': 'desire',
 'respect': 'admiration',
 'worry': 'fear',
 'pain': 'sadness',
 'lov': 'love',
 'apprehension': 'nervousness',
 'tragedy': 'sadness',
 'aid': 'relief',
 'love': 'joy',
 'confidence': 'optimism',
 'realization': 'realization',
 'praise': 'admiration',
 'anxiety': 'sadness',
 'interest': 'curiosity',
 'confusion': 'nervousness'}

Add to the dataset

df["tag2tag"] = df["Tags"].apply(lambda x: {cleaned_tag2tag[e] for e in x if e in cleaned_tag2tag})

#### 1.2.4. Use RoBERTa model to predict mapping of tags

tags = []
for e in tqdm(all_ds_tags):
    tags.append(pipe(e, truncation=True, top_k=None))

postpr = [{tag["label"] for tag in quote if tag["score"] > 0.1 and tag["label"] in my_set} for quote in tags]
tag_to_tags = {tag: tags for tag, tags in zip(all_ds_tags, postpr) if len(tags) != 0}

df["tag2tag_by_model"] = df["Tags"].apply(lambda x: {tt for e in x if e in tag_to_tags for tt in tag_to_tags[e] if tt in my_set})

df

#### 1.2.5. Combine all tags
Combine all tags into one column and save the dataset.

df["all_tags"] = df.apply(lambda x: x.model_tags.union(x.ds_tags).union(x.tag2tag).union(x.tag2tag_by_model), axis=1)

# make str
df["all_tags"] = df["all_tags"].apply(lambda x: ",".join(x))

df[['Quote', 'all_tags']].to_csv('../data/quotes_labeled_reddit.csv', index=False)

### 1.3. Add tags from the semeval dataset to the quotes dataset
It is necessary to add tags from the go emotions dataset to the quotes dataset, because we will use the model fine-tuned on go emotions dataset.

Here we will use the [twitter-roberta-base-emotion-multilabel-latest](https://huggingface.co/cardiffnlp/twitter-roberta-base-emotion-multilabel-latest) model.

#### 1.3.1. Use RoBERTa model to predict tags on quote text
Get the tags from the model (with threshold = 0.2) and filter them using cleaned tag set.

from transformers import pipeline
pipe = pipeline("text-classification", model="cardiffnlp/twitter-roberta-base-emotion-multilabel-latest")

my_set = set(pipe.model.config.label2id.keys())

tags = pipe(list(df["Quote"]), truncation=True, top_k=None)

df["model_tags"] = [{tag["label"] for tag in quote if tag["score"] > 0.4 and tag["label"] in my_set} for quote in tags]

#### 1.3.2. Add tags from the quotes dataset intersection with the go emotions dataset

df["ds_tags"] = df["Tags"].apply(lambda x: my_set.intersection(x))

df

#### 1.3.3. Use `fasttext` embeddings to find similar tags

import fasttext
from sklearn.metrics.pairwise import cosine_similarity

ft = fasttext.load_model('../cc.en.300.bin')

my_set_embed = [ft.get_word_vector(e) for e in my_set]

from tqdm.notebook import tqdm

all_ds_tags = set(df["Tags"].explode().value_counts().keys())
tag_to_tag = {}

for e in tqdm(all_ds_tags):
    for w, emb in zip(my_set, my_set_embed):
        if cosine_similarity([ft.get_word_vector(e)], [emb])[0][0] > 0.5:
            tag_to_tag[e] = w

Clean pairs of tags by hand.

tag_to_tag

cleaned_tag2tag = {'regret': 'sadness',
 'pleasure': 'joy',
 'lov': 'love',
 'passion': 'love',
 'sadness': 'pessimism',
 'happiness': 'optimism',
 'delight': 'surprise',
 'expectation': 'anticipation',
 'agony': 'sadness',
 'tragedy': 'sadness',
 'selfishness': 'anger',
 'self-confidence': 'optimism',
 'confidence': 'optimism',
 'pain': 'sadness',
 'laughter': 'joy',
 'apprehension': 'anticipation',
 'hatred': 'disgust',
 'jealousy': 'disgust',
 'anxiety': 'sadness',
 'excitement': 'optimism',
 'gratitude': 'joy',
 'love': 'love',
 'admire': 'love',
 'trust': 'trust',
 'amazement': 'joy',
 'pride': 'joy',
 'danger': 'fear',
 'hate': 'love',
 'rage': 'disgust',
 'shame': 'disgust',
 'worry': 'fear',
 'loneliness': 'sadness'}

Add to the dataset

df["tag2tag"] = df["Tags"].apply(lambda x: {cleaned_tag2tag[e] for e in x if e in cleaned_tag2tag})

#### 1.3.4. Use RoBERTa model to predict mapping of tags

tags = []
for e in tqdm(all_ds_tags):
    tags.append(pipe(e, truncation=True, top_k=None))

postpr = [{tag["label"] for tag in quote if tag["score"] > 0.5 and tag["label"] in my_set} for quote in tags]
tag_to_tags = {tag: tags for tag, tags in zip(all_ds_tags, postpr) if len(tags) != 0}

df["tag2tag_by_model"] = df["Tags"].apply(lambda x: {tt for e in x if e in tag_to_tags for tt in tag_to_tags[e] if tt in my_set})

#### 1.3.5. Combine all tags
Combine all tags into one column and save the dataset.

df["all_tags"] = df.apply(lambda x: x.model_tags.union(x.ds_tags).union(x.tag2tag).union(x.tag2tag_by_model), axis=1)

# make str
df["all_tags"] = df["all_tags"].apply(lambda x: ",".join(x))

df[['Quote', 'all_tags']].to_csv('../data/quotes_labeled_twitter.csv', index=False)

df

## 2. More Quotes by Parsing Some Sources

TODO: if needed
