==> notebooks/3.1-rectools-models.ipynb <==
# Experiments with models from RecTools

Generate mock data

import pandas as pd
import numpy as np
from datetime import timedelta

# Set a random seed for reproducibility
np.random.seed(42)

# Generate a DataFrame with 100 rows
data = []
start_time = pd.Timestamp.now()
time_step = timedelta(minutes=1)

for i in range(100):
    array1 = np.random.randint(-10, 10, size=(2,))
    array2 = np.random.randint(-10, 10, size=(2,))
    label = np.random.choice([-1, 1], size=(1,))[0]
    timestamp = start_time + i * time_step
    data.append([i, i+100, label, timestamp, array1.tolist(), array2.tolist()])
    # data.append([i, i+100, label, timestamp])
    # data.append([array1.tolist(), array2.tolist(), label, timestamp])

# Create the DataFrame
columns = ['User', 'Item', 'weight', 'Timestamp', 'Array1', 'Array2']
df = pd.DataFrame(data, columns=columns)
df.to_csv('../data/temp.csv')
# Display the DataFrame
print(df.info())

## Reading data into RecTools format

Loading labeled datasets and interactions dataset annotated by GPT-4.

from rectools import Columns
from rectools.dataset import Dataset
import pandas as pd
from datetime import timedelta


quotes_labeled_reddit = pd.read_csv('../data/quotes_labeled_reddit.csv', index_col=0)
quotes_labeled_reddit = quotes_labeled_reddit.replace('\r\n','\n', regex=True)

diaries_labeled_reddit = pd.read_csv('../data/diaries_labeled_reddit.csv', index_col=0)

raw_interactions = pd.read_csv('../data/diaries_quotes_emb_twitter_interactions.csv')

Construction of RecTools format dataset. It involves making feature vectores out of emotion label scores for user and item, assuming timestamps with 1 minute difference between each interaction.

rectools_data = []
rectools_diary_dense_features = []
rectools_quote_dense_features = []
rectools_diary_dense_fnames = None
rectools_quote_dense_fnames = None
diaries_index = set()
quotes_index = set()
start_time = pd.Timestamp.now()
time_step = timedelta(minutes=1)
for i, row in raw_interactions.iterrows():
    quote = quotes_labeled_reddit[quotes_labeled_reddit['Text'] == row['Quote']]
    quote_index = int(quote.index[0])
    quote_features = quote.drop('Text', axis=1).iloc[0].to_list()

    diary = diaries_labeled_reddit[diaries_labeled_reddit['Text'] == row['Text']]
    diary_index = int(diary.index[0])
    diary_features = diary.drop('Text', axis=1).iloc[0].to_list()

    timestamp = start_time + i * time_step
    if rectools_diary_dense_fnames is None:
        quote_fnames = quote.columns.to_list()[1:]
        diary_fnames = diary.columns.to_list()[1:]
        rectools_diary_dense_fnames = ['id'] + diary_fnames
        rectools_quote_dense_fnames = ['id'] + quote_fnames
    rectools_data.append([diary_index, quote_index, 2. if row['Interaction'] == 1 else 0.5, timestamp])
    if diary_index not in diaries_index:
        diaries_index.add(diary_index)
        rectools_diary_dense_features.append([diary_index] + diary_features)
    if quote_index not in quotes_index:
        quotes_index.add(quote_index)
        rectools_quote_dense_features.append([quote_index] + quote_features)

rectools_df = pd.DataFrame(rectools_data, columns=[*Columns.Interactions])
user_features_df = pd.DataFrame(rectools_diary_dense_features, columns=rectools_diary_dense_fnames)#.set_index('id')
item_features_df = pd.DataFrame(rectools_quote_dense_features, columns=rectools_quote_dense_fnames)#.set_index('id')
# user_features_df = user_features_df.astype('double')
# item_features_df = item_features_df.astype('double')
rectools_ds = Dataset.construct(rectools_df,
                            user_features_df=user_features_df,
                            cat_user_features=rectools_diary_dense_fnames,
                            item_features_df=item_features_df,
                            cat_item_features=rectools_quote_dense_fnames,
                            make_dense_user_features=True,
                            make_dense_item_features=True,
                            )
rectools_ds

### Train-test dataset splitting

from rectools.model_selection import RandomSplitter


splitter = RandomSplitter(test_fold_frac=0.2, random_state=42, n_splits=1, filter_cold_users=False,
                    filter_cold_items=False, filter_already_seen=False)
for train_ids, test_ids, _ in splitter.split(rectools_ds.interactions):
    print(train_ids, test_ids)

rectools_df_train = rectools_df.iloc[train_ids]
train_user_ids = rectools_df_train[Columns.User].to_list()
train_item_ids = rectools_df_train[Columns.Item].to_list()
user_features_df_train = user_features_df.loc[user_features_df['id'].isin(train_user_ids)]
item_features_df_train = item_features_df.loc[item_features_df['id'].isin(train_item_ids)]
rectools_ds_train = Dataset.construct(rectools_df_train,
                            user_features_df=user_features_df_train,
                            cat_user_features=rectools_diary_dense_fnames,
                            item_features_df=item_features_df_train,
                            cat_item_features=rectools_quote_dense_fnames,
                            make_dense_user_features=True,
                            make_dense_item_features=True,
                            )
rectools_ds_train

rectools_df_test = rectools_df.iloc[test_ids]
test_user_ids = rectools_df_test[Columns.User].to_list()
test_item_ids = rectools_df_test[Columns.Item].to_list()
user_features_df_test = user_features_df.loc[user_features_df['id'].isin(test_user_ids)]
item_features_df_test = item_features_df.loc[item_features_df['id'].isin(test_item_ids)]
rectools_ds_test = Dataset.construct(rectools_df_test,
                            user_features_df=user_features_df_test,
                            cat_user_features=rectools_diary_dense_fnames,
                            item_features_df=item_features_df_test,
                            cat_item_features=rectools_quote_dense_fnames,
                            make_dense_user_features=True,
                            make_dense_item_features=True,
                            )
rectools_ds_test

## Fit-predict

Making a recommedation table by using 5 different models, namely:

1. **PopularModel (Popularity-Based):**
   - Recommends top-k popular items for each user based on overall popularity.
2. **RandomModel (Random Recommendation):**
   - Provides random recommendations by selecting k items randomly for each user.
3. **PureSVDModel (Matrix Factorization):**
   - Utilizes Singular Value Decomposition for matrix factorization to recommend top-k items.
4. **AlternatingLeastSquaresModel (ALS):**
   - Collaborative filtering model using implicit feedback, recommending top-k items.
5. **LightFM (Hybrid Model):**
   - Hybrid model combining collaborative and content-based signals to recommend top-k items.

The choice of $k=10$ ensures a consistent evaluation metric, representing the number of recommendations each model provides for users in the recommendation tables.

from rectools.models import PopularModel

model = PopularModel()
model.fit(rectools_ds_train)
recos = model.recommend(
    users=rectools_df_train[Columns.User].unique(),
    dataset=rectools_ds_train,
    k=10,
    filter_viewed=True,
)

recos

from rectools.models import RandomModel


model_rnd = RandomModel()
model_rnd.fit(rectools_ds_train)
recos_rnd = model_rnd.recommend(
    users=rectools_df_train[Columns.User].unique(),
    dataset=rectools_ds_train,
    k=10,
    filter_viewed=True,
)

recos_rnd

from rectools.models import PureSVDModel


model_svd = PureSVDModel()
model_svd.fit(rectools_ds_train)
recos_svd = model.recommend(
    users=rectools_df_train[Columns.User].unique(),
    dataset=rectools_ds_train,
    k=10,
    filter_viewed=True,
)

recos_svd

from rectools.models import ImplicitALSWrapperModel
from implicit.als import AlternatingLeastSquares


model_als = ImplicitALSWrapperModel(
        AlternatingLeastSquares(
            factors=64,
            regularization=0.01,
            alpha=1,
            random_state=2023,
            use_gpu=False,
            iterations=15
            )
        )
model_als.fit(rectools_ds_train)
recos_als = model_als.recommend(
    users=rectools_df_train[Columns.User].unique(),
    dataset=rectools_ds_train,
    k=10,
    filter_viewed=True,
)

recos_als

from rectools.models import LightFMWrapperModel
from lightfm import LightFM


model_lfm = LightFMWrapperModel(
        model=LightFM(no_components = 30)
        )
model_lfm.fit(rectools_ds_train)
recos_lfm = model_lfm.recommend(
    users=rectools_df_train[Columns.User].unique(),
    dataset=rectools_ds_train,
    k=10,
    filter_viewed=True,
)

recos_lfm

## Evaluation

On the evalation stage we compute the following metrics:

1. **NDCG (Normalized Discounted Cumulative Gain):**
   - Measures the effectiveness of ranked recommendations. A higher NDCG score indicates better-ordered recommendations. Calculated for the top-k recommendations.
2. **Accuracy:**
   - Represents the proportion of correctly predicted recommendations among the total recommendations. Accuracy is computed for the top-k recommendations.
3. **IntraListDiversity:**
   - Quantifies the diversity of items within a recommended list. A higher score indicates a more diverse set of recommendations. Evaluated for the top-k recommendations.
4. **MAP (Mean Average Precision):**
   - Computes the average precision across different levels of recall. MAP is calculated for the top-k recommendations.
5. **MCC (Matthews Correlation Coefficient):**
   - Measures the correlation between actual and predicted recommendations. The MCC score is computed for the top-k recommendations.
6. **MRR (Mean Reciprocal Rank):**
   - Calculates the average of the reciprocal ranks of the first correct recommendation. MRR is evaluated based on the top-k recommendations.
7. **MeanInvUserFreq (Mean Inverse User Frequency):**
   - Measures the diversity of recommended items across users. A lower score indicates higher diversity. Computed for the top-k recommendations.
8. **Precision:**
   - Reflects the accuracy of positive recommendations among the total recommended items. Precision is calculated for the top-k recommendations.
9. **Recall:**
   - Measures the ability to retrieve relevant items among all relevant items. Recall is computed for the top-k recommendations.
10. **Serendipity:**
    - Assesses the novelty or unexpectedness of recommendations. Higher serendipity scores indicate more surprising recommendations. Evaluated for the top-k recommendations.

By consistently choosing $k=10$ for all metrics, the evaluation process ensures a standardized and meaningful comparison of recommendation performance across different models.

import warnings
warnings.filterwarnings("ignore")

from rectools.metrics import NDCG, Accuracy, IntraListDiversity, MAP, MCC, MRR, MeanInvUserFreq, Precision, Recall, Serendipity
from rectools.metrics.distances import PairwiseHammingDistanceCalculator
from sklearn.model_selection import train_test_split


k = 10

ndcg = NDCG(k=k, log_base=3)
acc = Accuracy(k=k)
mmap = MAP(k=k)
ild = IntraListDiversity(k=k, distance_calculator=PairwiseHammingDistanceCalculator(item_features_df_test))
mcc = MCC(k=k)
mrr = MRR(k=k)
miuf = MeanInvUserFreq(k=k)
prc = Precision(k=k)
rec = Recall(k=k)
srd = Serendipity(k=k)

interactions, prev_interactions = train_test_split(rectools_df_test[[Columns.User, Columns.Item]], train_size=0.2, random_state=42)  # Assumption
# interactions = rectools_df_test[[Columns.User, Columns.Item]]
# prev_interactions = rectools_df_test[[Columns.User, Columns.Item]]  # TODO: define
catalog = item_features_df_test['id'].unique()  # Collection of unique item ids that could be used for recommendations  

def compute_metrics(reco, model_name):
    print(model_name)
    print("NDCG: ", ndcg.calc(reco=reco, interactions=interactions))
    print("Accuracy: ", acc.calc(reco=reco, interactions=interactions, catalog=catalog))
    print("IntraListDiversity: ", ild.calc(reco=reco))
    print('MAP: ', mmap.calc(reco=reco, interactions=interactions))
    print('MCC: ', mcc.calc(reco=reco, interactions=interactions, catalog=catalog))
    print('MRR: ', mrr.calc(reco=reco, interactions=interactions))
    print('MeanInvUserFreq: ', miuf.calc(reco=reco, prev_interactions=prev_interactions))
    print('Precision: ', prc.calc(reco=reco, interactions=interactions))
    print('Recall: ', rec.calc(reco=reco, interactions=interactions))
    print('Serendipity: ', srd.calc(reco=reco, interactions=interactions, prev_interactions=prev_interactions, catalog=catalog))
    print()

compute_metrics(recos, 'PopularModel')
compute_metrics(recos_rnd, 'RandomModel')
compute_metrics(recos_svd, 'PureSVDModel')
compute_metrics(recos_als, 'AlternatingLeastSquaresModel')
compute_metrics(recos_lfm, 'LightFM')
